fix: resolve 6 compilation errors from marathon edits
1. TotalsAggregationEngine: div by Long -> toDouble()
2. ReceiptLifecycleCoordinator: added missing catch for return try block
3. RecurringLifecycleCoordinator: added Error branch in when exhaustiveness
4. WorkerSpecScheduler: ExistingWorkPolicy.UPDATE -> KEEP (UPDATE not available)
5. PrivacyBlockedCard: added CLOUD_AI_RECEIPT_OCR to when exhaustiveness

+++ app/src/main/java/com/yourname/expensetracker/domain/analytics/TotalsAggregationEngine.kt
@@ -553,7 +553,7 @@ class TotalsAggregationEngine @Inject constructor(
             ).toDays().coerceAtLeast(1)
         } ?: expenses.map { it.date }.distinct().size.coerceAtLeast(1)
         val average = if (expenses.isNotEmpty()) {
-            expenses.sumOf { it.normalizedAmount } / calendarDays
+            expenses.sumOf { it.normalizedAmount } / calendarDays.toDouble()
         } else 0.0
 
         // Group by day using the same dayKey format as [dayKey]
